10165: Remove the ability for irsPractitioners to multi-docket request access to case documents
