+++ web-client/src/presenter/actions/FileDocument/setDefaultFileDocumentFormValuesAction.test.ts
@@ -1,25 +1,18 @@
 import { GENERATION_TYPES } from '@web-client/getConstants';
-import { ROLES } from '../../../../../shared/src/business/entities/EntityConstants';
-import { allowExternalConsolidatedGroupFilingHelper } from '@web-client/presenter/computeds/allowExternalConsolidatedGroupFilingHelper';
 import { applicationContextForClient as applicationContext } from '@web-client/test/createClientTestApplicationContext';
+import {
+  petitionerUser,
+  privatePractitionerUser,
+} from '@shared/test/mockUsers';
 import { presenter } from '../../presenter-mock';
 import { runAction } from '@web-client/presenter/test.cerebral';
 import { setDefaultFileDocumentFormValuesAction } from './setDefaultFileDocumentFormValuesAction';
-jest.mock(
-  '@web-client/presenter/computeds/allowExternalConsolidatedGroupFilingHelper',
-);
 
 describe('setDefaultFileDocumentFormValuesAction', () => {
   presenter.providers.applicationContext = applicationContext;
-  const mockUserId = '082093a4-20e0-4a82-9d89-e24108535216';
 
   it('sets up the form with default values when allowExternalConsolidatedGroupFilingHelper returns true', async () => {
-    applicationContext.getCurrentUser.mockReturnValue({
-      role: ROLES.privatePractitioner,
-      userId: mockUserId,
-    });
-
-    allowExternalConsolidatedGroupFilingHelper.mockReturnValue(true);
+    applicationContext.getCurrentUser.mockReturnValue(privatePractitionerUser);
 
     const result = await runAction(setDefaultFileDocumentFormValuesAction, {
       modules: { presenter },
@@ -41,11 +34,7 @@ describe('setDefaultFileDocumentFormValuesAction', () => {
   });
 
   it('sets up the form with default values when allowExternalConsolidatedGroupFilingHelper returns false', async () => {
-    applicationContext.getCurrentUser.mockReturnValue({
-      role: ROLES.privatePractitioner,
-      userId: mockUserId,
-    });
-    allowExternalConsolidatedGroupFilingHelper.mockReturnValue(false);
+    applicationContext.getCurrentUser.mockReturnValue(privatePractitionerUser);
 
     const result = await runAction(setDefaultFileDocumentFormValuesAction, {
       modules: { presenter },
@@ -67,10 +56,7 @@ describe('setDefaultFileDocumentFormValuesAction', () => {
   });
 
   it('should set filersMap[userId] to true if the logged in user is a petitioner', async () => {
-    applicationContext.getCurrentUser.mockReturnValue({
-      role: ROLES.petitioner,
-      userId: mockUserId,
-    });
+    applicationContext.getCurrentUser.mockReturnValue(petitionerUser);
 
     const result = await runAction(setDefaultFileDocumentFormValuesAction, {
       modules: { presenter },
@@ -80,15 +66,12 @@ describe('setDefaultFileDocumentFormValuesAction', () => {
     });
 
     expect(result.state.form.filersMap).toEqual({
-      [mockUserId]: true,
+      [petitionerUser.userId]: true,
     });
   });
 
   it('should default te generationType to manual', async () => {
-    applicationContext.getCurrentUser.mockReturnValue({
-      role: ROLES.petitioner,
-      userId: mockUserId,
-    });
+    applicationContext.getCurrentUser.mockReturnValue(petitionerUser);
 
     const result = await runAction(setDefaultFileDocumentFormValuesAction, {
       modules: { presenter },
