+++ web-client/src/presenter/state.ts
@@ -15,7 +15,6 @@ import { addTrialSessionInformationHelper } from './computeds/TrialSession/addTr
 import { advancedDocumentSearchHelper } from './computeds/AdvancedSearch/advancedDocumentSearchHelper';
 import { advancedSearchHelper } from './computeds/AdvancedSearch/advancedSearchHelper';
 import { alertHelper } from './computeds/alertHelper';
-import { allowExternalConsolidatedGroupFilingHelper } from './computeds/allowExternalConsolidatedGroupFilingHelper';
 import { appInstanceManagerHelper } from './computeds/appInstanceManagerHelper';
 import { applyStampFormHelper } from './computeds/applyStampFormHelper';
 import { batchDownloadHelper } from './computeds/batchDownloadHelper';
@@ -150,7 +149,6 @@ export const computeds = {
   advancedDocumentSearchHelper,
   advancedSearchHelper,
   alertHelper,
-  allowExternalConsolidatedGroupFilingHelper,
   appInstanceManagerHelper,
   applyStampFormHelper,
   batchDownloadHelper,
